% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipschema-package.R
\docType{package}
\name{pipschema-package}
\alias{pipschema}
\alias{pipschema-package}
\title{pipschema: Identify and Validate Schemas in PIP API Data}
\description{
Tools to identify, validate, and document data schemas used by the Poverty and Inequality Platform (PIP) API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PIP-Technical-Team/pipschema}
  \item Report bugs at \url{https://github.com/PIP-Technical-Team/pipschema/issues}
}

}
\author{
\strong{Maintainer}: Rossana Tatulli \email{rtatulli@worldbank.org}

}
\keyword{internal}
